comment-rate-service/requirements.txt
django
djangorestframework
requests
PyMySQL
scikit-learn


cryptography
